refactor: split manager media endpoints into dedicated router
